modify: getThumbnail util added to PopualarContents, CommunityPostList, ByMangoPostList, and PostCard.tsx
