+++ src/pages/detail/components/detailBody/skeleton/detailBodySkeleton.css
@@ -1,7 +1,6 @@
 .detailBody-skeleton-container {
   width: 100%;
   height: 400px;
-  /* padding: 20px; */
   margin-top: 40px;
   display: flex;
   flex-direction: column;
