+++ src/pages/detail/components/detailHeader/DetailHeader.tsx
@@ -1,13 +1,13 @@
-import { useState } from 'react';
-import 'swiper/css';
-import { SwiperClass, SwiperSlide } from 'swiper/react';
 import swipeLeft from 'assets/icons/swipeLeft.png';
 import swipeRight from 'assets/icons/swipeRight.png';
 import useSwiperNavigation from 'hooks/useSwiperNavigation';
+import { useState } from 'react';
+import 'swiper/css';
+import { SwiperClass, SwiperSlide } from 'swiper/react';
 import { FoundDetailPostProps } from 'types/PostType';
 
-import St from './style';
 import { convertToKor } from 'pages/write/components/common/lists';
+import St from './style';
 
 function DetailHeader({ foundDetailPost }: FoundDetailPostProps) {
   const [swiperInstance, setSwiperInstance] = useState<SwiperClass | null>(null);
@@ -15,6 +15,7 @@ function DetailHeader({ foundDetailPost }: FoundDetailPostProps) {
 
   const coverImages = foundDetailPost.coverImages;
 
+  // swiper custom hook
   const { goNext, goPrev } = useSwiperNavigation({
     swiperInstance,
     currentIndex,
@@ -27,36 +28,34 @@ function DetailHeader({ foundDetailPost }: FoundDetailPostProps) {
   };
 
   return (
-    <>
-      <St.CoverContainer>
-        {coverImages.length > 0 ? (
-          <>
-            <St.StyledSwiper onSwiper={setSwiperInstance} onSlideChange={handleSlideChange} className="mySwiper">
-              {coverImages.map((image, idx) => {
-                return (
-                  <SwiperSlide key={idx}>
-                    <img src={image.url} alt="cover" />
-                    <St.Gradient></St.Gradient>
-                    <St.PostHeaderInfo $noImage={false}>
-                      <span>{convertToKor(foundDetailPost.category)}</span>
-                      <h2>{foundDetailPost.title}</h2>
-                    </St.PostHeaderInfo>
-                  </SwiperSlide>
-                );
-              })}
-            </St.StyledSwiper>
-            <St.NavigationButtonContainer>
-              <div onClick={goPrev}>{currentIndex > 0 && <img src={swipeLeft} alt="Previous" />}</div>
-              <div onClick={goNext}>{currentIndex < coverImages.length - 1 && <img src={swipeRight} alt="Next" />}</div>
-            </St.NavigationButtonContainer>
-          </>
-        ) : (
-          <St.NoImage>
-            <St.PostHeaderInfo $noImage={true}>{foundDetailPost.title}</St.PostHeaderInfo>
-          </St.NoImage>
-        )}
-      </St.CoverContainer>
-    </>
+    <St.CoverContainer>
+      {coverImages.length > 0 ? (
+        <>
+          <St.StyledSwiper onSwiper={setSwiperInstance} onSlideChange={handleSlideChange} className="mySwiper">
+            {coverImages.map((image, idx) => {
+              return (
+                <SwiperSlide key={idx}>
+                  <img src={image.url} alt="cover" />
+                  <St.Gradient></St.Gradient>
+                  <St.PostHeaderInfo $noImage={false}>
+                    <span>{convertToKor(foundDetailPost.category)}</span>
+                    <h2>{foundDetailPost.title}</h2>
+                  </St.PostHeaderInfo>
+                </SwiperSlide>
+              );
+            })}
+          </St.StyledSwiper>
+          <St.NavigationButtonContainer>
+            <div onClick={goPrev}>{currentIndex > 0 && <img src={swipeLeft} alt="Previous" />}</div>
+            <div onClick={goNext}>{currentIndex < coverImages.length - 1 && <img src={swipeRight} alt="Next" />}</div>
+          </St.NavigationButtonContainer>
+        </>
+      ) : (
+        <St.NoImage>
+          <St.PostHeaderInfo $noImage={true}>{foundDetailPost.title}</St.PostHeaderInfo>
+        </St.NoImage>
+      )}
+    </St.CoverContainer>
   );
 }
