+++ src/pages/write/components/imageUpload/imageHandler.ts
@@ -7,7 +7,6 @@ const imageHandler = (quillRef: React.RefObject<ReactQuill>) => {
   input.setAttribute('type', 'file');
   input.setAttribute('accept', 'image/*');
   input.click();
-  //console.log(19950811);
 
   input.addEventListener('change', async () => {
     if (!quillRef.current || !input.files) return;
