+++ src/recoil/posts.ts
@@ -1,5 +1,4 @@
 import { atom } from 'recoil';
-// import { Category } from '../components/viewAll/ViewAllBody';
 import { Category } from 'types/PostListType';
 import { IsEditingPostProps, PostInputType } from 'types/PostType';
 import { POST } from './keys';
